pdf-generator/requirements.txt
pypdf>=4.0
reportlab>=4.0
python-docx>=1.0
Pillow>=10.0
numpy>=1.24
